# docs/openapi/all.yaml
openapi: "3.0.3"
info:
  title: User API
  version: "1.0"
servers:
  - url: http://localhost:3000/api
    description: unsecure
  - url: https://localhost:3000/api
paths:
...
  /roles:
    get:
      tags:
        - Role Management
      summary: Fetch roles
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: name
          schema:
            type: string
          description: find name like
        - in: query
          name: size
          schema:
            type: integer
          description: limit the size (default 10)
        - in: query
          name: page
          schema:
            type: integer
          description: page number (default 1)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FetchRoleListResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /roles/{id}:
    get:
      tags:
        - Role Management
      summary: Fetch role by id
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/GetRoleResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        size:
          type: integer
          example: 10
        total:
          type: integer
          example: 100
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            title:
              type: string
            message:
              type: string
...
    GetRoleResponse:
      type: "object"
      properties:
        id:
          type: "integer"
        name:
          type: "string"
        idname:
          type: "string"
          example: "1 - Superadmin"
    FetchRoleListResponse:
      type: "object"
      properties:
        data:
          type: "array"
          items:
            $ref: "#/components/schemas/GetRoleResponse"
        meta:
          type: "object"
          properties:
            namelike:
              type: "string"
              nullable: true
            pagination:
              $ref: "#/components/schemas/Pagination"
...
  securitySchemes:
    BearerAuth: # Logical name for the scheme
      type: http
      scheme: bearer
      bearerFormat: JWT # Optional, tells the client it's a JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: auth_session